+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## 2025.04.04
+
+### Added
+
+- Designed abstract class for playlist component
+
+### Updated
+
+- Added length() to kernel interface
+- Changed remove() to removeFront()
+- Added randomSong() to secondary interface
+
 ## 2025.03.09
 
 ### Added